// source/types/constants/ErrorCode.ts
export enum ErrorCode {
    NOT_FOUND = "NOT_FOUND",
    BAD_REQUEST = "BAD_REQUEST",
    DUPLICATE_ENTRY = "DUPLICATE_ENTRY",
    NOT_UNIQUE = "NOT_UNIQUE",
    FACEBOOK_ID_NOT_FOUND = "FACEBOOK_ID_NOT_FOUND",
    PHONE_NOT_UNIQUE = "PHONE_NOT_UNIQUE",
    WEB_ADDRESS_NOT_UNIQUE = "WEB_ADDRESS_NOT_UNIQUE",
    EMAIL_NOT_UNIQUE = "EMAIL_NOT_UNIQUE",
    EMAIL_NOT_VALID = "EMAIL_NOT_VALID",
    FACEBOOK_ID_NOT_UNIQUE = "FACEBOOK_ID_NOT_UNIQUE",

    INTERNAL_ERROR = "INTERNAL_ERROR",
    AUTH_ERROR = "AUTH_ERROR",
    ABORTED_ERROR = "ABORTED",

    WRONG_EXTENSION_ERROR = "WRONG_EXTENSION_ERROR",
    SIZE_EXCEEDED_ERROR = "SIZE_EXCEEDED_ERROR",
}